still not working to send the logs to Grafana. Every log is not working for some reason

+++ src/metrics.js
@@ -100,7 +100,7 @@ class MetricTracker {
                     sum: {
                       dataPoints: [
                         {
-                          asFloat: metricValue,
+                          asInt: metricValue,
                           timeUnixNano: Date.now() * 1000000,
                           attributes: [],
                         },
